- Changing the repositories to work with the Random.me API - Adding Room integration with the Pagination Adapter and the BondaryCallback to have an "offline-first" UX - Changing some dependencies to work with the latest version of AS

+++ app/src/main/java/com/kimboo/mvvmkotlin/extensions/DataSource.kt
@@ -5,7 +5,7 @@ import android.support.annotation.IntDef
 class DataSource<MODEL>(var model: MODEL?, @param:DataSourceState @field:DataSourceState @get:DataSourceState var state: Int) {
 
     @Retention()
-    @IntDef(SOURCE_HTTP_NOT_MODIFIED.toLong(), SOURCE_HTTP_SUCCESS.toLong(), SOURCE_DATABASE.toLong(), SOURCE_MIXED.toLong())
+    @IntDef(SOURCE_HTTP_NOT_MODIFIED, SOURCE_HTTP_SUCCESS, SOURCE_DATABASE, SOURCE_MIXED)
     annotation class DataSourceState
 
     companion object {
